removed redundant logging calls and moved health check from CMS to reader so its always available
