+++ zebedee-reader/src/main/java/com/github/onsdigital/zebedee/reader/api/endpoint/Health.java
@@ -1,4 +1,4 @@
-package com.github.onsdigital.zebedee.api;
+package com.github.onsdigital.zebedee.reader.api.endpoint;
 
 import com.github.davidcarboni.restolino.framework.Api;
 import org.apache.http.HttpStatus;
@@ -6,6 +6,8 @@ import org.apache.http.HttpStatus;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.GET;
+
+import static com.github.onsdigital.zebedee.logging.ReaderLogger.info;
 
 /**
  * health endpoint always returns 200 OK.
